Unset client ID when stopped
When the client invokes `stop`, it is signalling an explicit intent to
disconnect. Therefore, subsequent connections should not be considered
'resume' events, but complete 'begin' events in their own right.

By unsetting the client ID, multiple `run` invocations each behave
identically (each sending a new 'begin' event to the server), and the
'resume' event is reserved for only unexpected re-connections.